requirements.txt
Flask==3.0.0
gunicorn==21.2.0
reportlab==4.0.7
PyPDF2==4.0.1
python-docx==0.8.11